Problem Statement:
Given a list of coordinates, find those coordinates that fall into the
given threshold and assign them the smallest or largest value as passed
in the parameters and remove them.
First line contains the coordinate list.
Second line gives the threshold and the third line takes in a boolean(0
or 1) where 0 means use the smallest value and 1 means use the largest
value


Example 1:
[341, 373, 375, 450, 453, 480, 560, 561, 703]
5
0
Result: [341, 373, 450, 480, 560, 703]

Example 2:
[452, 455, 480, 483, 502, 504, 600, 630, 720]
4
1
Result: [455, 483, 504, 600, 630, 720]